Add CI check for deterministic codegen
