+++ .github/workflows/ci.yml
@@ -110,6 +110,8 @@ jobs:
           runner: ubuntu-latest
         - action: check-tools
           runner: smithy_ubuntu-latest_8-core
+        - action: check-deterministic-codegen
+          runner: smithy_ubuntu-latest_8-core
     steps:
     - uses: actions/checkout@v3
       with:
