+++ tools/ci-scripts/check-deterministic-codegen
@@ -8,14 +8,10 @@ set -eux
 cd smithy-rs
 
 if [[ $# -ne 1 ]]; then
-    echo "Usage: generate-codegen-diff <base-revision>"
+    echo "Usage: check-deterministic-codegen"
     exit 1
 fi
 
 # Override version commit hash to prevent unnecessary diffs
 export SMITHY_RS_VERSION_COMMIT_HASH_OVERRIDE=ci
-base_revision="$1"
-./tools/ci-scripts/codgen-diff/codegen-diff-revisions.py . "${base_revision}"
-
-mv tmp-codegen-diff/bot-message ../artifacts/bot-message-codegen-diff
-mv tmp-codegen-diff ../artifacts/codegen-diff
+./tools/ci-scripts/codgen-diff/check-deterministic-codegen.py .
