+++ tools/ci-scripts/codegen-diff/__init__.py
@@ -1,0 +1,2 @@
+#  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#  SPDX-License-Identifier: Apache-2.0
